+++ viadot/task_utils.py
@@ -8,6 +8,8 @@ from typing import List, Literal
 
 import pandas as pd
 import prefect
+import pyarrow as pa
+import pyarrow.dataset as ds
 from prefect import task
 from prefect.storage import Git
 from prefect.utilities import logging
@@ -303,6 +305,41 @@ def df_converts_bytes_to_int(df: pd.DataFrame) -> pd.DataFrame:
     logger = prefect.context.get("logger")
     logger.info("Converting bytes in dataframe columns to list of integers")
     return df.applymap(lambda x: list(map(int, x)) if isinstance(x, bytes) else x)
+
+
+@task
+def df_to_dataset(
+    df: pd.DataFrame, partitioning_flavor="hive", format="parquet", **kwargs
+) -> None:
+    """
+    Use `pyarrow.dataset.write_to_dataset()` to write from a pandas DataFrame to a dataset.
+    This enables several data lake-specific optimizations such as parallel writes, partitioning,
+    and file size (via `max_rows_per_file` parameter).
+
+    Args:
+        df (pd.DataFrame): The pandas DataFrame to write.
+        partitioning_flavor (str, optional): The partitioning flavor to use. Defaults to "hive".
+        format (str, optional): The dataset format. Defaults to 'parquet'.
+        kwargs: Keyword arguments to be passed to `write_to_dataset()`. See
+        https://arrow.apache.org/docs/python/generated/pyarrow.dataset.write_dataset.html.
+
+    Examples:
+        table = pa.Table.from_pandas(df_contact)
+        base_dir = "/home/viadot/contact"
+        partition_cols = ["updated_at_year", "updated_at_month", "updated_at_day"]
+
+        df_to_dataset(
+            data=table,
+            base_dir=base_dir,
+            partitioning=partition_cols,
+            existing_data_behavior='overwrite_or_ignore',
+            max_rows_per_file=100_000
+        )
+    """
+    table = pa.Table.from_pandas(df)
+    ds.write_dataset(
+        data=table, partitioning_flavor=partitioning_flavor, format=format, **kwargs
+    )
 
 
 class Git(Git):